chore(dev-server): don't rely on prebuilding
